requirements.txt
# ================================
# requirements.txt
# ================================
torch
torchvision
matplotlib
Pillow